requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
selenium>=4.0.0
cloudscraper>=1.2.71
fake-useragent>=1.4.0
langchain==0.3.1
langchain-openai>=0.3.1
langchain-community>=0.3.1
langchain-core>=0.3.1
playwright>=1.40.0
openai==1.99.3
python-dotenv==1.0.0
numpy>=1.21.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
